Added support for routing middleware
- The implementation is a simple chain of handlers that are called
one after the other in the order they were defined. The api is simpler
than the ring model where you need to return a function and call the next handler.
Here this is taken care of by the library. The user supplies a sequence of
functions. In blocking mode, the return value of each function is the argument
of the next function. In async mode, the value emitted by calling `respond` is the
argument of the next function.
- Refactored RouteDescriptor to take a collection of handlers
- The user functions are reified into Handler<RoutingContext> rather than
Function<RoutingContext, ?>, which makes a lot more sense. It also eliminates
the need for RingHandler.java and BlockingRingHandler.java.
- Instead of RING_REQUEST_FIELD and RING_RESPONSE_FIELD the result of each handler
is added to LAST_HANDLER_RESPONSE_FIELD.
- Updated Spec
- Fixed tests